+++ changelog.md
@@ -4,6 +4,7 @@
 - Fix formatting/ToString of char.
 - Fix issue when using a local variable inside an object initializer.
 - Fix Template.Evaluate if used with Liquid.
+- Fix argument out of the range when using TemplateContext.IndentWithInclude.
 
 ## 3.0.3 (29 Nov 2020)
 - Fix parsing of comment with a code exit when parsing mode is script only.